requirements.txt
streamlit
python-dotenv
google-generativeai
txtai
PyPDF2
youtube_transcript_api
pathlib